+++ modules/core/Collections/views/index.php
@@ -21,8 +21,8 @@
         <div class="uk-navbar-flip">
             <div class="uk-navbar-content">
                 <div class="uk-button-group">
-                    <button class="uk-button" data-ng-class="mode=='list' ? 'uk-button-primary':''" data-ng-click="(mode='list')" title="@lang('List mode')" data-uk-tooltip="{pos:'bottom'}"><i class="uk-icon-th"></i></button>
-                    <button class="uk-button" data-ng-class="mode=='table' ? 'uk-button-primary':''" data-ng-click="(mode='table')" title="@lang('Table mode')" data-uk-tooltip="{pos:'bottom'}"><i class="uk-icon-list-alt"></i></button>
+                    <button class="uk-button" data-ng-class="mode=='list' ? 'uk-button-primary':''" data-ng-click="setListMode('list')" title="@lang('List mode')" data-uk-tooltip="{pos:'bottom'}"><i class="uk-icon-th"></i></button>
+                    <button class="uk-button" data-ng-class="mode=='table' ? 'uk-button-primary':''" data-ng-click="setListMode('table')" title="@lang('Table mode')" data-uk-tooltip="{pos:'bottom'}"><i class="uk-icon-list-alt"></i></button>
                 </div>
             </div>
         </div>
@@ -52,7 +52,7 @@
     </div>
 
     <div class="app-panel" data-ng-if="collections && collections.length && mode=='table'">
-        <table class="uk-table">
+        <table class="uk-table uk-table-striped">
             <thead>
                 <tr>
                     <th>@lang('Collection')</th>
@@ -65,9 +65,13 @@
                         <a href="@route('/collections/collection')/@@ collection._id @@">@@ collection.name @@</a>
                     </td>
                     <td align="right">
-                        @hasaccess?("Collections", 'manage.collections')
-                        <a class="uk-text-danger" data-ng-click="remove($index, collection)" href="#" title="@lang('Delete collection')" data-uk-tooltip="{pos:'bottom'}"><i class="uk-icon-minus-circle"></i></a>
-                        @end
+                        <ul class="uk-subnav uk-subnav-line">
+                            <li><a href="@route('/collections/entries')/@@ collection._id @@" title="@lang('Show entries')" data-uk-tooltip="{pos:'bottom'}"><i class="uk-icon-bars"></i></a></li>
+                            <li><a href="@route('/collections/entry')/@@ collection._id @@" title="@lang('Create new entry')" data-uk-tooltip="{pos:'bottom'}"><i class="uk-icon-plus-circle"></i></a></li>
+                            @hasaccess?("Collections", 'manage.collections')
+                            <li><a class="uk-text-danger" data-ng-click="remove($index, collection)" href="#" title="@lang('Delete collection')" data-uk-tooltip="{pos:'bottom'}"><i class="uk-icon-minus-circle"></i></a></li>
+                            @end
+                        </ul>
                     </td>
                 </tr>
             </tbody>
@@ -79,7 +83,7 @@
         <p class="uk-text-large">
             @lang('You don\'t have any collections created.')
         </p>
-        
+
         @hasaccess?("Collections", 'manage.collections')
         <a href="@route('/collections/collection')" class="uk-button uk-button-success uk-button-large">@lang('Create a collection')</a>
         @end