+++ modules/core/Forms/views/index.php
@@ -20,8 +20,8 @@
         <div class="uk-navbar-flip">
             <div class="uk-navbar-content">
                 <div class="uk-button-group">
-                    <button class="uk-button" data-ng-class="mode=='list' ? 'uk-button-primary':''" data-ng-click="(mode='list')" title="@lang('List mode')" data-uk-tooltip="{pos:'bottom'}"><i class="uk-icon-th"></i></button>
-                    <button class="uk-button" data-ng-class="mode=='table' ? 'uk-button-primary':''" data-ng-click="(mode='table')" title="@lang('Table mode')" data-uk-tooltip="{pos:'bottom'}"><i class="uk-icon-list-alt"></i></button>
+                    <button class="uk-button" data-ng-class="mode=='list' ? 'uk-button-primary':''" data-ng-click="setListMode('list')" title="@lang('List mode')" data-uk-tooltip="{pos:'bottom'}"><i class="uk-icon-th"></i></button>
+                    <button class="uk-button" data-ng-class="mode=='table' ? 'uk-button-primary':''" data-ng-click="setListMode('table')" title="@lang('Table mode')" data-uk-tooltip="{pos:'bottom'}"><i class="uk-icon-list-alt"></i></button>
                 </div>
             </div>
         </div>
@@ -51,7 +51,7 @@
     </div>
 
     <div class="app-panel" data-ng-if="forms && forms.length && mode=='table'">
-        <table class="uk-table">
+        <table class="uk-table uk-table-striped">
             <thead>
                 <tr>
                     <th>@lang('Form')</th>
@@ -78,7 +78,7 @@
         <p class="uk-text-large">
             @lang('You don\'t have any forms created.')
         </p>
-        
+
         @hasaccess?("Forms", 'manage.forms')
         <a href="@route('/forms/form')" class="uk-button uk-button-success uk-button-large">@lang('Create a form')</a>
         @end