+++ modules/core/Regions/assets/js/index.js
@@ -2,7 +2,7 @@
 
     App.module.controller("regions", function($scope, $rootScope, $http, $timeout){
 
-        $scope.mode = 'list';
+        $scope.mode = App.storage.get("cockpit.view.listmode", 'list');
 
         $http.post(App.route("/api/regions/find"), {}).success(function(data){
 
@@ -33,6 +33,12 @@
 
         $scope.matchName = function(name) {
             return (name && name.indexOf($scope.filter) !== -1);
+        };
+
+        $scope.setListMode = function(mode) {
+            $scope.mode = mode;
+
+            App.storage.set("cockpit.view.listmode", mode);
         };
     });